+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# BCN 0.4.0
+
+* Remove 'direct' method for now
+* Improve `verbose` argument in `bcn` (now an integer in (0, 1, 2, 3))
+
 # BCN 0.3.1
 
 * Remove some dependencies to speed up package's startup